Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

There is an associated adjustment in sepolicy to enable access to /dev/kmsg. The *default* action with this CL is to log to *both* places. A change from "1" to "" for the getprop is all that is needed to change this. I have confirmed, to my pleasure, that the logs go to both accurately; and if I <ghasp> restart logd it re-inherits from dmesg the audit logs before continuing to be the primary receiver of the content. This later side effect resolves the issue that if <heaven forbid> logd crashes that the audit logs get some permanence.

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1