Introduce the notion of zygote ABIs.
Each zygote sets a property that defines the list of ABIs
it supports. The union of this list is "ro.cpu.abilist", which
is the list of all ABIs the device supports.

When the frameworks decides to launch an app, it will look
at the highest matching ABI and delegate to the correct zygote.

This information could be added to the target config makefiles
(which would allow us to make it read only) but it seems cleaner
to put all zygote configuration information in the same file.

bug: 13647418

Change-Id: I874cd9515a2633c462558dc62e9403e347bf4532