Update patch set 1
Patch Set 1:

Yes, libcorkscrew is deprecated and replaced by libunwind.

It's also not clear under what circumstances a map indicates it's readable, but it isn't.  Is it possible that the map cache is actually wrong and that's the reason for the crash?

If you can duplicate this failure, please re-run with libunwind and tell us if it continues to be a problem and I can tell you where a similar patch might need to be added.

Patch-set: 1