utils: fix compiler warnings
Recent changes in RefBase.h introduced two compiler warnings, which
show up if the system/core/include path is included explicitly. If
the file is not included explicitly, it gets included through the
-isystem option which suppresses the warnings.

The warnings are from incorrect initialization order and from a
missing virtual destructor. This patch fixes those.

Change-Id: Iff76a655eb8bd547adfe994c7315a005e98aed41
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Mingwei Shi <mingwei.shi@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>