Charger: Charging-status to be checked to remain on
At present, online property is checked to decide if charging in progress.
In some scenario, this may not be propagated to charger properly, or
online could be 1, but charging not in progress. In such sceanrio, it doesn't
make sense to remain in charger-mode if not charging.
Hence, adding charging-status also to confirm about Charging/Discharging.

Change-Id: Ib8fe5c798b48a260e929e0432e0b1a3c7dc6351f
Signed-off-by: Pavan Kumar S <pavan.kumar.s@intel.com>
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Signed-off-by: Yong Yao <yong.yao@intel.com>