Update patch set 1
Patch Set 1:

Greg had pointed out combinations of flags can result in unexpected results. I did a test, the global removal of -Wno-unused did not have the same effect as adding -Wunused locally, which captured a few more problems here and there (?)

I will opt for cleaning the code (as some of this leads to simplification moving forward), we can use a separate CL to enforce if we change our minds about making it permanent.

Patch-set: 1