Update patch set 1
Patch Set 1: Code-Review+1

Thanks for the pair of CLs. This takes a lot of pressure off of the 64-bit porting efforts, while still complying with a win_sdk build. I will meander through the side-effects on the builds with my branches, but I have not set myself up to verify the win_sdk functionality.

Patch-set: 1