Update patch set 1
Patch Set 1:

This needs to be applied more widely, include/log/log.h is yet another example; everywhere we see __attribute__.*printf/scanf if an audit is to be explored; however we could deal with this on a case-by-case basis since we are dealing with a subset of components? bionic will also be affected. Tip of the iceberg?

Patch-set: 1