Update patch set 2
Patch Set 2:

That's not what I meant.

I meant that the surface flinger and zygote are always started and stopped together, so why not an ultra thin wrapper around those ?

Anyway, I'm still against adding these types of wrappers. They add more complexity and levels of indirection. For example
(1) the wrapper will need a SIGCHLD handler to allow it to kill itself when one of the zygotes die to make restart rules work.

(2) The wrapper will need to map environment variables init sets (as is the convention when you set socket rules) and convert them into command line arguments for the child process.

The nice way of doing this would be allow init to bring down all services in a given class via the system property observer (instead of creating thin wrappers every time services need to be started or stopped together...)

Patch-set: 2