Update patch set 4
Patch Set 4:

Thanks for reviewing. I tested this new patch on a Haswell device and observed the proper pci/ symlinks, and also ran it on a Baytrail device with MMC and observed that platform/ symlinks are still created as before.

Patch-set: 4
Reviewer: Gerrit User 1011333 <1011333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0