Update patch set 1
Patch Set 1:

This is going to lead to inconsistent behavior.  If the first adb command you happen to run has '-a', it will spawn a server listening on all interfaces.  Otherwise, if the server already exists, the '-a' will be silently ignored.

For special behavior like this, it seems easier to manually spawn the adb server with "adb server nodaemon -a &"

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0