watchdogd: drop root privileges
The watchdogd daemon is part of the init binary code which has the
sticky bit set.  Hence the watchdod daemon is started as a root
process whatever what is specified in the init.rc file.

CTS tests require a justification for all the process that are running
as a root process.  This daemon does not need such a privilege.  This
patch makes it drop its privilege and migrate to the system uid.

Change-Id: If4726ebdc31b1f5afface263d40b2d822228697b
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>