watchdogd: drop root privileges
CTS tests require a justification for all the process that are running
with root privileges.  This daemon does not need those privileges.
This patch makes it drop them by migrating to the system uid.

The watchdogd daemon is part of the init code it is even a symbolic
link to the /init binary file which is root.  As a consequence the
watchdogd needs to be started as a root process before it drops its
privleges by itself.

Change-Id: If4726ebdc31b1f5afface263d40b2d822228697b
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>