Update patch set 1
Patch Set 1:

x86_64 syscall time is 2us on much hotter processors than we are using with Android, so syscall overhead is a fixed and architecture/processor specific entity that *all* user space developers should be cognizant of the impact. In almost all cases, 'it is what it is'. Since Manta is worse than Volantis I would say we do not have as much of a leg to stand on to optimize the kernel mechanisms. Real Time user-space applications expect that syscalls can result in many thousands of user-space instructions of time. For instance the audio folks stay away from syscalls as if they are the plague.

Patch-set: 1