Add setservicecon built-in command and call it.
Add a setservicecon built-in command to set the default
SELinux security context for services that have neither
an explicit seclabel option nor an automatic domain
transition defined in policy.  Call the command from
early-init in the init.rc to set the default context to
use the service domain defined in an external/sepolicy
change with the same Change-Id.

This ensures that no service is left running in init's domain.

Change-Id: I01ce683748919e53beabe6906d6074e6c00d4f2d
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>