Update patch set 1
Patch Set 1:

I was thinking an UID & PID match. I do believe (with 90% of my being) that UID is the right point of security delineation.

Regardless the experience thus far is that the PIDs have never (I know, never say never) rolled over within the lifetime of the logs, even when I had the spam filter running (which would hold on to less noisy application's data for considerably longer). The logger does in fact, within the Statistics handling track when an application disappears from its PID, and records the applications comm, so we can correlate this and use it to our advantage with regards to a PID-only solution too.

NB: We do allow the user to set the log lengths up to 16M (https://android-review.googlesource.com/#/c/94015) though so I am of course only speaking of the default 256KB logs size when it comes to log rollover.

Patch-set: 1