Update patch set 1
Patch Set 1:

UID only filtering feels better to me than UID+PID filtering.

Imagine a process fork/execing a binary it can't modify, and wanting to cleanup the output of that binary. Because the child process would have a different PID than it's parent, it wouldn't be able to cleanup the logged data. UID+PID also feels problematic in that, if an app is killed and restarted, it won't be able to clear log data from it's previous run.

Longer term, UID + selinux_domain, IMHO, is how I'd prefer to filter. This allows us to distinguish between, for example, vold and netd. Both of those processes run with UID=0, but vold shouldn't be able to clear netd's logs.

With the current change, UID=0 is all powerful, but I think it's worth trying to figure out how we can avoid that special case.

Anyway, this change is perfectly good to me. I'm very happy with it. Just suggesting ideas for future possible directions.

Patch-set: 1