Update patch set 1
Patch Set 1:

I would very much hope that developers are not running with setenforce=0 unless they have found/suspect that they are being blocked by an SELinux bug, and waiting for its resolution.  SELinux is a part of the environment, and running with it disabled could lead one to act on false assumptions.

Patch-set: 1