Update patch set 1
Patch Set 1:

Long reply:

I discussed this earlier with Nick and we're moving toward a world wherein our SElinux policy is a reflection of our security model. SELinux now provides enough additional protection as to be necessary on its own, rather than acting only as a fail-safe for our existing protection mechanisms.  One of the primary benefits of using a MAC system is the ability to comprehensively evaluate our security policy in one place.  Another benefit is that the protection provided by SElinux is fundamentally different from that previously offered, and in this case, it provides a much better mechanism of protection than that which we had previously.  We have already taken measures, e.g. https://android-review.googlesource.com/#/c/96642/, to make sure that there is no easy way to disable selinux.  We are also mandating it as part of compatibility, and plan to do everything we can to make sure it is running properly.

I definitely do not think that we should maintain the property_service only for those situations when SELinux is not involved.  Either it should be maintained in addition to SELinux, or not at all.  Doing a half-solution in this case is likely to lead to mistakes.  For this particular change, Nick has verified that we lose nothing (we are no less strict) by going to SELinux only, and since SELinux is now an essential part of our security model, which bypassing would be a much larger problem, I'm ok with removing this check.

Patch-set: 1
Reviewer: Gerrit User 1026058 <1026058@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0