Stop and start netd explicitly for "adb shell stop/start".
netd is supposed to be restarted when the zygote is restarted (see the
"onrestart" section for "service zygote" in init.zygote*.rc). However, in
practice, it doesn't always work, since "stop; start" isn't always recognized as
a "restart".

It seems to depend on timing. On a Nexus 5 with an "eng" build, a delay less
than 200ms between stop and start seems to be treated as a restart. Anything
more, and netd doesn't get restarted.

Bug: 15855807
Change-Id: I387fe86600f4a2862abc3a05a2ef9a1e7374e21d