Update patch set 1
Patch Set 1:

on this subject, note that cp(1) is from upstream, but contains a bunch of hacks. (and i'm about to add another one so we can copy /proc files that claim to be zero-length...)

i think grep(1) is 99% upstream; the only hackery left being to remove the compression support (?).

Patch-set: 1