Update patch set 2
Patch Set 2: Code-Review+2

I'm fine with this if you want it, but it seems out of place.  If netd can survive a stop/start after your other change, why do you want to stop and start it here?  I believe the onrestart in init.rc is to deal with cases where the native daemon has gotten into a bad state and is what is causing the runtime restart, not to tie the lifecycles of zygote and netd together.

Patch-set: 2
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2