Fix for adb progress reporting (-p) on windows.
fstat() was being called with a pseudo handle (lfd)
that was not an actual windows file handle.
Changed the code to use stat() instead and pass the
path to the file.
Also, changed the logic so that if stat() fails the
progress will not be shown, but the push operation
will not be aborted.

Change-Id: Ia55809ae50a82c4fee2dc9bcbf4f980ed366a8b2