Update patch set 3
Patch Set 3:

Did "include external/libcxx/libcxx.mk" not work ? That make file appears to have been written for target modules, but it should work for host modules as well.

I'm going to approve this change as is because if the security guys want this cherry-picked to LMP-dev, I don't want to pull in this new dependency and destabilize things.

Patch-set: 3
Reviewer: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0