Add UseUTF8ForNamesEncoding public method to ziparchive.
Add new public method to allow checking if UTF-8 should be used to encode entry
names. If not then they will be encoded in IBM PC character encoding.

All entries in an archive will be encoded with the same encoding. Either
UTF-8 or IBM PC. Some archives use both of these encodings. In such case
we will use the one which is more used in an archive. This can cause a problem
if names in an archive contain characters which are represented differently
in UTF-8 and IBM PC.

Bug: 16162465
Change-Id: I6f81b540496c7b9f4b042fd223e2bb4911e960ff