Update patch set 2
Patch Set 2:

It's just a result of our decision to trade support for archives with multiple encodings for performance. Ideally we would like all archives to be UTF-8 only. The ones which use only IBM PC are fine too. The one which mix encodings but use only English are fine too.

This is not a regression. It worked this way previously.

Patch-set: 2
Label: Verified=0