Update patch set 2
Patch Set 2:

I have had my 15 minutes of fame ;->. I used a substantial majority of the adjustments, so I would prefer to apply an update to _this_ CL retaining Pierre's originating authorship if that is OK? This does not disconnect me from commit, your or peer review, ownership or future maintenance. Is that OK? If not, I probably can also adjust authorship within the context of this CL's history.

Patch-set: 2