Update patch set 1
Patch Set 1:

Thanks for the comment!

First let me use here as a discussion place.
 > You can use su if you want to escalate to root.
But there is no su command when build with the user variant.

This problem comes to me because that the CTS test needs to be run with the user build to have 100% pass rate for the selinux related tests.
And if we run the CTS test with the user mode, we want to run other tests with the same user mode build too, not want to run tests with different build.
And since we can not change to the su domain to run the tests, some test will fail because seclabel u:r:shell:s0 does not have permission to do something like writing proc file or reading/writing device file.


Any suggestions for this test cases?

Patch-set: 1
Reviewer: Gerrit User 1041813 <1041813@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0