Update patch set 1
Patch Set 1:

And what I consider is to set console service with the superuser permission and with su security context.

And make su domain available when ro.debuggable is 1.(now available when build in the eng or userdebug variant)

But I am not sure what problems will be caused here, and do not know if there is any reason to set console service with shell user.

Patch-set: 1