Update patch set 3
Patch Set 3:

I see you removed atom arch from the cache header file. If that means that Android does not support the old atom arch anymore, then we should update the memsetX functions to slm versions as well (based on the slm memset in bionic).
The reason I kept the filenames was because I had plans to do a SSSE3 version for Atom and an SSE4 version for Silvermont.

Patch-set: 3