Switch adb to libusbhost.
Using strace(1) while switching adb to epoll(2), I noticed that it was doing a
lot of work every second. Turns out that rather than using libusbhost (which
uses inotify) it was polling using readdir(3). Switching to libusbhost makes
the trace very quiet, and everything still seems to work.

It looks like a lot more of this code could be replaced by calls to libusbhost,
but I don't really know enough about adb or libusbhost to want to dig much
deeper. This seems like a step in the right direction at least.

Bug: 17588403
Change-Id: Ib814253cc186022b15c82af27d68bc4aaed98a10