Turn a shutdown request into reboot if configured
So far this workaround is mainly to reboot to charge-only mode
on some platforms while actually the bootloader can do anything
once it gets a "power off" request from a reboot. In this case,
we expect the bootloader to decide if it should boot to charge-
only mode by checking the charger status.

Not all of the PMICs support to boot device automatically into
charge-only mode after shutdown when charger is connected. The
below is some explanation of why we make a change in system/
core project.

() There is no "reboot" HAL for vendor-specific change. Reboot
and shutdown are based on syscall provided by Linux kernel.

() A change in upper layers in Java will miss some corner cases.
e.g: recovery mode where Java services are not launched.

() The healthd is the component that provides status of charger
connection status between kernel and Android framework.

() The init is the component that performs the actual reboot in
Android system. We also keep some native non-android daemons
from being effected in this way, as long as they call the reboot
APIs from any lower levels than init's.

() Putting the workaround in kernel doesn't sound better:
  - We will have to hack the standard reboot routines in kernel.
  - It is also difficult to differentiate shutdown requests
    from user space.
  - Acquire charger(s) status in reboot sub-system will actually
    be a re-implementation of what healthd in user space does.

In a short, this change tries to resolve the issue in a least
destructive way. Two properties are invented. One is to enable
or disable overriding shutdown request. Another holds connection
status of charger(s).

Change-Id: Ie96fb58ecb9a4e8ffe7e73e590483b5323e40b74
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>