adb shell <cmd> sets exit code correctly.
This requires us to change adbd to prepend the executing PID to the
returned output when sent an rshell: command, and the adb client then
strips this PID off the output, and uses a new retval:<PID> command to
ask for the exit code associated with the PID. This is necessary because
it doesn't seem to be possible to append the exit code to the output
that is sent back with the shell: command.

This seems to withstand stress testing, only failing the supplied
test script when either:
  - there has been an ADB protocol error (printed out), or
  - ADB has disconnected
due to the amount of connections being made anyway, which has been
observed regardless of this patch.

There is one issue: if we do adb shell dalvikvm in a way that causes it
to exit abnormally, (e.g., reporting a NullPointerException) this seems
to manifest as the subprocess being killed with a SIGHUP, which is
returned to the caller of adb shell as exit code 129. However, when
running the same command within an interactive shell on the device
itself (i.e., if you do adb shell, and then dalvikvm) just returns 1, as
expected. We welcome any advice on what might be happening here.

Change-Id: I8591ca67a2b4ceb62167b1986f53bf651e8c5828
Signed-off-by: Stephen Kyle <stephen.kyle@arm.com>