libsparse: Use ssize_t/off_t for file length/offset
Because the image to be flashed might exceed 2G, even 4G, which will
be out of the range of int or uint type. So we need to use 64-bit int to
indicate the file length. With the new build option "-D_FILE_OFFSET_BITS=64",
ssize_t and off_t will be defined with int64_t.

Change-Id: I3e91dcc623ebdb0934831d9ff7a8880868e93411
Signed-off-by: Xiao-Feng Lei <xiaofeng.lei.bj@gmail.com>