libsparse: read huge file data in smaller part
This patch fixes the issue that the huge image (exceeding 4G) can't
be sparsed correctly. The issue was caused by 1) fastboot was built
as a 32-bit ELF file on Linux/Mac, and it fails to mmap 4G file (actually
a little less than 4G file is also failed to be mapped), and 2) fastboot
tried to allocate a memory with size of the image on Windows, and failed.
To resovle this issue, on Linux/Mac, the image is mmapped in 3G blocks,
and it's read out in max 2G block on Windows.

Change-Id: I602db630a90e3421b465a1e7e3679d5e3e78f32a
Signed-off-by: Xiao-Feng Lei <xiaofeng.lei.bj@gmail.com>