Update patch set 3
Patch Set 3:

Would someone like to approve this again?

This is like the last attempt, except the SMP=0 case was removed, and as a result the various MEMORY_ORDER macros could be removed, cleaning up (and correcting!) the code.

The autoverifier -1 appears to have been bogus, due to an unrelated intermittent failure.  I'm still testing, but it seems the rest of the tree has advanced to the point that that, with the just committed mingw symbolic link, this actually works.  I'd like to check this in today rather than just before I leave on vacation.  It's still conceivable there's another symbolic link missing someplace or the like.

Patch-set: 3