Init: Add a runtime check for the TARGET_CPU_SMP build parameter.
This patch adds a runtime check for the SMP build parameter. This needs to be
set for multi-core systems and unset for single core ones. A multi-core system
with TARGET_CPU_SMP set to false will fail when running init.

In theory this should prevent debugging issues related to non-SMP builds running
on SMP systems.

Change-Id: Idb75dad4457fbcced6a558c25b82f425ae41ecda
Signed-off-by: Serban Constantinescu <serban.constantinescu@arm.com>