Disable vendor id filtering for OS X.
Companion patch to change I2bf583a27f9a3f98ef006ea62b906a4f89960507.

Due to the use of USB class 0xff it's not possible to register a listener
which uses the OS methods for filtering notifications, so this patch
also introduces a check in the notification listener which checks the
class, subclass, and protocol supplied in notifications.

Change-Id: I380cc80f96addc5057660efb2179b35f9378886a