fastboot: possibility to print usb download progress
Can be used in to see the progress of a usb download and for parsing
it in a GUI program for showing a progress bar respectively.

It comes with a slight usability-drawback regarding the download
of sparse images. In this case, the progress of every chunk will
be printed separately.

Change-Id: Id267eeeddd4decb29f07c43d28f3d369aa40e30a
Signed-off-by: Florian Bäuerle <florian.bae@gmail.com>