bootchart: fix bootchart can not be triggered problem
bootchart use the file in data partition to decide
if it should start collect data for bootchart.
but the data partition will be mounted by the mount_all command
in the "on fs" section, and it will be only added into the action
queue when command "trigger fs" is execute, and it is after
the position of bootchart_init action which is added into
the action queue just after late_init section.
This causes the bootchart will not be started since
the start flag file is not there before it's partition is mounted.

This change makes bootchart_init a builtin command of init,
and add the init.bootchart.rc file to make it executed as the last command
of "on fs" section, and the position of bootchart_init can be moved
to any place in init.rc series file if you like.

Change-Id: Ia74aa34ca5b785f51fcffdd383075a549b2a99d9
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>