bootchart: fix bootchart can not be triggered problem
bootchart use the file in data partition to decide
if it should start collect data for bootchart.
but the data partition will be mounted by the mount_all
command in the "on fs" section, and it will be only added
into the action queue when command "trigger fs" is execute,
and it is after the position of bootchart_init action which is
added into the action queue just after late_init section.
This causes the bootchart will not be started since the start flag
file is not there before it's partition is mounted.

This change makes bootchart_init a builtin command of init,
and make it executed as the first command of "on post-fs" section
which will be trigger after "on fs" section. This will let bootchart
can collect as much data as possible.

Change-Id: Ia74aa34ca5b785f51fcffdd383075a549b2a99d9
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>