Bump adb server version
With the recent work on adb server side code (e.g. the removal of the USB
VID whitelist) it's probably a good idea to bump the server version number.

This number is also shown in the client side banner (which shows
major.minor.server version numbers), so it'll be easier to determine if
any bug reports relate to pre- or post- server side changes.

Change-Id: Ic550d72905ae406bba68a96534978838a2289689