Update patch set 1
Patch Set 1:

I've not had any issues personally, but given the work which has been done it'll be easier to determine if any bug reports could be related to the recent work with a version bump. We could bump the adb minor version instead of the server protocol version, but it's been at 0 for so long I was wondering if the changes are significant enough.

Patch-set: 1
Reviewer: Gerrit User 1010485 <1010485@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0