init: Add support for virtior-ports symlinks
Virtio serial devices are used in virtual environments to provide a
simple form of paravirtualized char communication with a backend,
typically from the host system running the Android emulator.

Such serial devices, however, show up per default in /dev/vport%up%u,
and there is no stable ABI for their numbering, so userspace has no
established way of tying a port to a backend.

The virtio serial protocol (and the kernel driver) does support naming
ports, and those names show up in the device's sysfs directory.  It is
slow an unpractical for userspace services that use this ports to have
to iterate through sysfs, so instead create symlinks for the ports using
the port name, based on the existing infrastructure in init.

Change-Id: Id6a65953cc8461e49ae1574558c288ca6f8bb12c
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>