Update patch set 1
Patch Set 1:

Yes, saw Traian's CL. Have commented with an alternate to his #1 problem that is more efficient and does not result in memory pressure, thanking him for the #2 solution.

On pstore, we have https://android-review.googlesource.com/#/c/118720 for the kernel side with the intent of pushing upstream. For user space, we have prototypes we are in the middle of unit testing and should likely publish within the week in AOSP. This is all contingent on the team evaluating if we want this feature (the issue is that one _must_ consider kernel panics much more important than adjusting user space to move away from kernel sensitivity; a natural side effect of having user space mixed with kernel during triage).

Patch-set: 1