Update patch set 1
Patch Set 1:

(1 comment)

Drop message simple & debug (disabled in user builds). If the message has an easy-to-parse format (binary 'event'?), then logd statistics (typically enabled in userdebug and eng) can present a report? Just some ideas, not ready to head straight to implementation yet.

Kernel logger was less likely to report EBUSY, but it did happen. When we ran a spam pressure test, our goal was to be within range on both for dropped content; but the test was run _without_ a log reader running (my bad). We need to revise the test ...

Patch-set: 1