liblog: Add local LOGGER_FIFO frontend
Low-latency logging utilizing a local consumer worker thread and
circular buffer selected using android_set_log_frontend(LOGGER_FIFO).
Additional benchmarks to check the effects of the algorithm. The
algorithm works regardless of caller's scheduler policy.

There are occurrences of system calls or latency under certain circumstances:

1) multiple threads writing to the logs when contended introduces syscalls
2) first call on a lapse of 100ms (overhead < 20us) introduces wakeup syscall
3) first call after selecting frontend performs a fair amount of initialization
4) latency is higher if CPU clockrate drops due to low CPU utilization

Otherwise calling latency of a LOG message is typically less than a few us
or a few 100 of ns. Latency improvements by a factor of 4 or more in benchmarks.
Algorithm should only be selected for native applications that suffer from
latency as this increases memory pressure and CPU utilization.

Please be advised that 10hz (100ms) logging rate is considered log spamming,
optimization is to deal with a burst of log messages. Latency is not the
time from logging to it appearing in the logs, it references the period
of time to push a log message into the FIFO.

Change-Id: If9b7fb25c4ee6dbaf9b7dd0d9993c12cab0af233